feat: added func getInfoByToken & logOut

+++ TccBackEnd/Domain/Interfaces/IAuthRepository.cs
@@ -7,7 +7,7 @@ public interface IAuthRepository
 {
   Task<Result<string>> Logar(LogarCredenciaisDto dto);
   Task<Result<string>> CadastrarUsuario(Usuario usuario);
-  Task<Result<string>> LogOut(long id);
+  Task<Result<string>> LogOut(int id);
   Task<Result<string>> ForgotPassword(string email);
   Task<Result<string>> VerificarEmail(string token);
 }
